Move SessionOverview Component
Putting this under session/overview with the other new session stuff now
that it is modernized.
