+++ packages/ilios-common/addon/components/session-details.hbs
@@ -8,7 +8,7 @@
   class="session-details"
   data-test-session-details
 >
-  <SessionOverview
+  <Session::Overview
     @session={{@session}}
     @editable={{@editable}}
   />
