+++ packages/ilios-common/addon/components/session-publicationcheck.hbs
@@ -2,7 +2,7 @@
   class="session-publicationcheck"
   data-test-session-publicationcheck
 >
-  <SessionOverview
+  <Session::Overview
     @session={{@session}}
     @hideCheckLink={{true}}
     @sessionTypes={{this.sessionTypes}}
