+++ packages/ilios-common/addon/components/session/overview.js
