+++ packages/ilios-common/addon/templates/session/copy.hbs
@@ -1,5 +1,5 @@
 <div class="session-details">
-  <SessionOverview
+  <Session::Overview
     @session={{@model}}
     @editable={{this.canUpdate}}
     @showCheckLink={{false}}
