--- packages/ilios-common/app/components/session-overview.js
@@ -1,1 +1,0 @@
-export { default } from 'ilios-common/components/session-overview';
