+++ packages/ilios-common/app/components/session/overview.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/components/session/overview';
